Restructured file architecture.
    The plugin's entry file is its bootstrap.  The intent of that
    file is to get the plugin up and running.  All of the other
    tasks should be abstracted into separate single responsibility
    files which focus on a single reason to exists.

    With this in mind, this commit divides up the code and
    re-organizes it into individual files.

    Why?  Why would you want to do this?

    1. Readability. It's hard to read a file that has a lot of code in
    it, especially when the code is not related or grouped.
    2. Improves being able to find what you are looking for, as all
    related code is grouped together.  For example, all of the admin
    bar functions are grouped in a file called src/admin/admin-bar.php.
    3. It's more maintainable.
    4. Compliance with the Single Responsibility Principle.
